chore(type): finalize state usage of campaign types enum
