+++ types/campaign.ts
@@ -129,3 +129,11 @@ export interface CombinedCampaignData {
   location: string | null;
   treasuryAddress: string | null;
 }
+
+export enum CampaignStatus {
+  DRAFT = 'DRAFT',
+  PENDING_APPROVAL = 'PENDING_APPROVAL',
+  ACTIVE = 'ACTIVE',
+  COMPLETED = 'COMPLETED',
+  FAILED = 'FAILED',
+}
